upload fee06cd6a4d46b5bfb0d4f6615a037854b3e46b5

+++ bmcv/test/test_vpss_convert_to.c
@@ -34,7 +34,7 @@ typedef struct border {
     u8 value_r;
     u8 value_g;
     u8 value_b;
-    u8 thickness;
+    u16 thickness;
 } border_t;
 
 typedef struct coverex_cfg {
